refactor: RestClient 를 /infrastructure 로 이동 및 인터페이스 추가
