+++ src/main/java/roomescape/application/support/TossPaymentService.java
@@ -1,0 +1,10 @@
+package roomescape.application.support;
+
+import roomescape.dto.request.TossPaymentConfirmDto;
+import roomescape.dto.response.TossPaymentConfirmResponseDto;
+
+public interface TossPaymentService {
+
+    TossPaymentConfirmResponseDto requestConfirmation(TossPaymentConfirmDto tossPaymentConfirmDto);
+
+}
