+++ src/main/java/roomescape/common/config/WebConfig.java
@@ -9,7 +9,7 @@ import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 import roomescape.presentation.interceptor.AdminAuthInterceptor;
 import roomescape.presentation.resvoler.LoginMemberArgumentResolver;
-import roomescape.presentation.support.TossPaymentWithRestClient;
+import roomescape.infrastructure.TossPaymentWithRestClient;
 
 @Configuration
 public class WebConfig implements WebMvcConfigurer {
