+++ src/main/java/roomescape/presentation/controller/ReservationTicketController.java
@@ -43,6 +43,7 @@ public class ReservationTicketController {
     @ResponseStatus(HttpStatus.CREATED)
     public ReservationTicketResponseDto addReservation(
             @RequestBody @Valid ReservationTicketRegisterDto reservationTicketRegisterDto, LoginMember loginMember) {
+
         return reservationTicketService.saveReservation(reservationTicketRegisterDto, loginMember);
     }
